+++ src/lib/auth-components/guilds/guild-card.svelte
@@ -59,7 +59,7 @@
 
 <form
 	method="post"
-	action="/?/checkIfJoinedAction"
+	action="/APIS?/checkIfJoinedAction"
 	enctype="multipart/form-data"
 	use:enhance={checkIfJoinedActionNews}
 >
