+++ src/lib/auth-components/learning-module/upload-module/upload-module-form.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import Loader from '$lib/general-components/loader.svelte';
 	import { fade } from 'svelte/transition';
 </script>
 
@@ -32,7 +31,7 @@
 	<button
 		class="bg-main w-full rounded-[10px] text-[14px] font-semibold py-[10px] px-[2px] flex items-center justify-center text-submain"
 	>
-		<Loader name="Create" txtColor="text-submain text-[14px]" />
+		Create
 	</button>
 
 	<a
