+++ src/routes/(authenticated)/+layout.svelte
@@ -3,6 +3,7 @@
 	import { getAuthState, getUserState, setAuthState, setUserState } from '$lib';
 	import ProjectJoinedContent from '$lib/auth-components/projects/join-project/project-joined-content.svelte';
 	import type { LayoutServerData } from './$types';
+	import { enhance } from '$app/forms';
 
 	export let data: LayoutServerData;
